+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-﻿<a name="0.9.2"></a>
+﻿<a name="0.9.3"></a>
+# 0.9.3
+
+* Added ability to include sub resources via the expand API query parameter
+** Orders can include Customer
+** Invoices can include Customer
+** Payments can include Fee Summary
+** Refunds can include Fee Summary
+
+<a name="0.9.2"></a>
 # 0.9.2
 
 * Renamed package and wording to be "data export" instead of "data exporter"